Add files via upload
# The Great Real Estate Data Challenge

**Background**
An investment fund that is planning to invest in properties at a number of locations. Using property sales over the past few years, the fund wants to identify properties with higher ROIs. The fund requires that the properties be segmented for easy analysis. This project will help the fund easily identify their target properties using advanced AI and Analytics.

**Dataset & Features**
We have received a dataset of real estate properties with locality, estimated price, and selling price for the past 13 years.

**Project Goal**
 The task is to identify the input features in the dataset and use them to predict the sale price of a property. After this modelling of input and output features, predict the sale price of all the properties in the test dataset. Once the sale prices for the test data are predicted, put these properties into 4 segments. These segments can be formed according to the gain. The gain is calculated based on the estimated price and predicted sale price (Gain = (Sale price - Estimated price)/100).

The results must be submitted as the segment level for each of the properties given in the test data.  For reference, the properties need to be segmented into the following 4 segments according to the gain calculated based on the predicted sale:-
- 0: Premium Properties
- 1: Valuable Properties
- 2: Standard Properties
- 3: Budget Properties